Reorganized MapView for ease of updates, introducing UI.Art package.
