+++ src/UI/Art/MapArtist.java
@@ -1,0 +1,9 @@
+package UI.Art;
+
+import java.awt.Graphics;
+
+public interface MapArtist {
+	public void drawMap(Graphics g);
+	public void drawCursor(Graphics g);
+	public void drawHighlights(Graphics g);
+}
