+++ src/UI/Art/MenuArtist.java
@@ -1,0 +1,9 @@
+package UI.Art;
+
+import java.awt.Graphics;
+
+import UI.GameMenu;
+
+public interface MenuArtist {
+	public void drawMenu(Graphics g);
+}
