+++ src/UI/Art/UnitArtist.java
@@ -1,0 +1,7 @@
+package UI.Art;
+
+import java.awt.Graphics;
+
+public interface UnitArtist {
+	public void drawUnits(Graphics g);
+}
